sendcommand/requirements.txt
cryptography==3.1.1
flaky==3.6.1
gcp-devrel-py-tools==0.0.15
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth==1.6.3
google-cloud-pubsub==1.0.0
google-cloud-iot==0.3.0
grpc-google-iam-v1==0.12.3
pyjwt
paho-mqtt==1.4.0
